Added Speech Spectrum Shaped Noise
We took an input of 30 different speech samples. These speech samples were randomized and added together to generate the speech spectrum shaped noise. The sampling frequency of this noise is 10000 Hz; same as the sampling frequency of the audio files of the single and concurrent vowels, which were down sampled from 100000 Hz to 10000 Hz.